services/controller/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
pyjwt==2.9.0
requests==2.32.3
